chore(soql): refactor lsp client code. Split files
