+++ packages/salesforcedx-vscode-soql/test/vscode-integration/testUtilities.ts
@@ -93,6 +93,7 @@ export const mockSObjects = [
         groupable: true,
         relationshipName: null,
         sortable: true,
+        type: 'id',
         updateable: false
       },
       {
